httpd: limit SSL certificate to 2 years if clock is accurate
Apple intends to start rejecting certificates that have a duration
longer than 2 years.  Still go with 10 years when clock ain't set,
to avoid certificates being already expired when generating it on
a firmware build that's older than two years.

We might possibly want to also check for expiration after ntp
is set (and generate a new certificate if necessary)?